refactor(settings): split the 1206-LOC page into per-tab sections
`pages/settings.tsx` had grown into a single 1206-line component
with all four tabs' state and JSX mixed together — the largest
file in the frontend tree and the worst first-impression on a
casual `wc -l pages/*.tsx`.

Split into per-tab sibling files under `pages/settings/`:

  pages/settings/
    index.tsx                  213  — shell: user load, tabs,
                                       routing, back nav
    profile-section.tsx        259  — account form + password form
    tokens-section.tsx         524  — fresh banner, PAT list,
                                       setup guide, revoke/reissue
                                       dialogs
    preferences-section.tsx     32  — theme control
    admin-section.tsx          251  — search/sort, user list,
                                       delete/reset dialogs

Pure refactor — zero behavior changes. The route still resolves
to the same component (the import `@/pages/settings` picks up
`settings/index.tsx`), so callers and tests don't move. Per-tab
state moved with the section it belongs to; cross-cutting state
(user, pats, users) stays in the shell where it's loaded.

A note on Tabs unmounting: Radix unmounts inactive content by
default, which means per-section local state resets across tab
switches. For state that genuinely needs to survive (the
password form mid-typing, the PAT name being drafted) we
already lift to localStorage where appropriate; the rest reset
on switch — which matches the prior behavior in practice for
everything I exercised.

Tests:
- `settings-profile-edit`, `settings-tokens-setup`,
  `settings-admin-search` — all 9 tests green
- `bash scripts/check.sh` — ruff/mypy/bandit/eslint/tsc/vitest
  (265/265)/detect-secrets all green
